requirements.txt
numpy==1.20.1
torchvision==0.9.0
psutil==5.8.0
faiss-gpu==1.7.2
tqdm==4.48.2
torch==1.8.0
Pillow==8.2.0
scikit_learn==0.24.1
torchscan==0.1.1
requests==2.26.0
timm==0.3.2
transformers==4.28.1
einops==0.4.1
